requirements.txt
streamlit
langchain
faiss-cpu
openai
google-generativeai
unstructured